Fix Func update and add uninstall
