+++ package.json
@@ -58,6 +58,7 @@
         "onCommand:azureFunctions.startStreamingLogs",
         "onCommand:azureFunctions.stopStreamingLogs",
         "onCommand:azureFunctions.deleteProxy",
+        "onCommand:azureFunctions.uninstallFuncCoreTools",
         "workspaceContains:local.settings.json",
         "workspaceContains:host.json",
         "onView:azureFunctionsExplorer"
@@ -232,6 +233,11 @@
             {
                 "command": "azureFunctions.deleteProxy",
                 "title": "%azFunc.deleteProxy%",
+                "category": "Azure Functions"
+            },
+            {
+                "command": "azureFunctions.uninstallFuncCoreTools",
+                "title": "%azFunc.uninstallFuncCoreTools%",
                 "category": "Azure Functions"
             }
         ],
