+++ src/commands/createNewProject/validateFuncCoreToolsInstalled.ts
@@ -6,12 +6,15 @@
 import * as opn from 'opn';
 import { MessageItem } from 'vscode';
 import { callWithTelemetryAndErrorHandling, DialogResponses, IActionContext } from 'vscode-azureextensionui';
-import { funcPackageName, PackageManager, Platform } from '../../constants';
+import { PackageManager } from '../../constants';
 import { ext } from '../../extensionVariables';
+import { getFuncPackageManager } from '../../funcCoreTools/getFuncPackageManager';
+import { installFuncCoreTools } from '../../funcCoreTools/installFuncCoreTools';
 import { localize } from '../../localize';
 import { getFuncExtensionSetting, updateGlobalSetting } from '../../ProjectSettings';
 import { cpUtils } from '../../utils/cpUtils';
 
+// Leaving this here for the purposes of review. I will move this to src/funcCoreTools/validateFuncCoreToolsInstalled.ts before merging
 export async function validateFuncCoreToolsInstalled(forcePrompt: boolean = false): Promise<boolean> {
     let input: MessageItem | undefined;
     let installed: boolean = false;
@@ -54,7 +57,7 @@ export async function validateFuncCoreToolsInstalled(forcePrompt: boolean = fals
 
                 if (input === install) {
                     // tslint:disable-next-line:no-non-null-assertion
-                    await attemptToInstallLatestFunctionRuntime(packageManager!);
+                    await installFuncCoreTools(packageManager!);
                     installed = true;
                 } else if (input === DialogResponses.dontWarnAgain) {
                     await updateGlobalSetting(settingKey, false);
@@ -77,66 +80,11 @@ export async function validateFuncCoreToolsInstalled(forcePrompt: boolean = fals
     return installed;
 }
 
-export async function attemptToInstallLatestFunctionRuntime(packageManager: PackageManager, runtimeVersion?: string): Promise<void> {
-    const v1: string = 'v1';
-    const v2: string = 'v2';
-    if (process.platform !== Platform.Windows) {
-        runtimeVersion = v2;
-    } else if (!runtimeVersion) {
-        const v1MsgItm: MessageItem = { title: v1 };
-        const v2MsgItm: MessageItem = { title: v2 };
-        runtimeVersion = (await ext.ui.showWarningMessage(localize('windowsVersion', 'Which version of the runtime do you want to install?'), v1MsgItm, v2MsgItm)).title;
-    }
-
-    ext.outputChannel.show();
-    switch (packageManager) {
-        case PackageManager.npm:
-            if (runtimeVersion === v1) {
-                await cpUtils.executeCommand(ext.outputChannel, undefined, 'npm', 'install', '-g', funcPackageName);
-            } else if (runtimeVersion === v2) {
-                await cpUtils.executeCommand(ext.outputChannel, undefined, 'npm', 'install', '-g', `${funcPackageName}@core`, '--unsafe-perm', 'true');
-            }
-            break;
-        case PackageManager.brew:
-            await cpUtils.executeCommand(ext.outputChannel, undefined, 'brew', 'tap', 'azure/functions');
-            await cpUtils.executeCommand(ext.outputChannel, undefined, 'brew', 'install', funcPackageName);
-            break;
-        default:
-            break;
-    }
-}
-
 export async function funcToolsInstalled(): Promise<boolean> {
     try {
         await cpUtils.executeCommand(undefined, undefined, 'func', '--version');
         return true;
     } catch (error) {
         return false;
-    }
-}
-
-export async function getFuncPackageManager(isFuncInstalled: boolean): Promise<PackageManager | undefined> {
-    switch (process.platform) {
-        case Platform.Linux:
-            // https://github.com/Microsoft/vscode-azurefunctions/issues/311
-            return undefined;
-        case Platform.MacOS:
-            try {
-                isFuncInstalled ?
-                    await cpUtils.executeCommand(undefined, undefined, 'brew', 'ls', funcPackageName) :
-                    await cpUtils.executeCommand(undefined, undefined, 'brew', '--version');
-                return PackageManager.brew;
-            } catch (error) {
-                // an error indicates no brew; continue to default, npm case
-            }
-        default:
-            try {
-                isFuncInstalled ?
-                    await cpUtils.executeCommand(undefined, undefined, 'npm', 'ls', '-g', funcPackageName) :
-                    await cpUtils.executeCommand(undefined, undefined, 'npm', '--version');
-                return PackageManager.npm;
-            } catch (error) {
-                return undefined;
-            }
     }
 }
